1/04/2025 - Home
Worked on the UI for the Login Page, making it more user friendly, it is having some problems with error code printing, but It can be resolved later. I learnt how the pack function works, and how to obtain data such as where the user is typing, which I used throughout the login function. It was a productive process today.